Category functionality - Read Desc
When a category is on the database, it will be displayed on the homepage per the brief request. Once that category is clicked on, it will redirect the user to the products page that will only show the products in that category that was clicked on.

This does not interfere with the products page at all, this will only show once you click the category name as if you were to click the products page in the navbar, nothing is there.

Soon will replace the text as images for frontend to align it with CSS and fancy designs.