Cart Quantity + Aligned All Tabs + Cart CSS + If Cart is Empty
Added functional CSS to the cart, added cart quantity button and everything performs as it should be. As well as all the tables are fixed, so the manage products table and manage user tables are not shifted one to the left as well as the new cart table that I've added will align well with the contents.

Added an IF function that can be used everywhere now if there are no products, or if there are no carts. In this case, I made it so if there are no products in the cart then the user will see nothing (no checkout, total amount) or anything that does not make sense. Instead, they will see "Your cart is currently Empty"