Added Wishlist Functionality including Wishlist page
CSS is already formatted because I reused my css for cart page yesterday.

Instead of adding products to the session like a carts page. You can now add products to the wishlist table, so the products actually save when logging out which is what wishlist intends to do.

Remove is a bit weird right now because apparently any logged in user has a user_id of null. Will soon fix.