Fixed User null error on wishlist
Now, the user_id links with the wishlist and is not null anymore.

This removes the ability of another user seeing another user's wish list despite the product being linked to the wrong user_id. This is by implementing authentication to the addwish button in maincontroller.

Made navbar consistent throughout pages too. Remove wishlist is next.