Inventory Management - Stock Indication Alert + Fix Index Html
Index HTML was messy and caused many things to lose centring, so I fixed that as well as repeating CSS code.

Products if low stock (under 10) will be displayed as 'low stock' visually for the user, as well as if it is in-stock (10 or over) and finally if there is no stock it will be displayed to the user.

If there is no stock then the user will not be able to add said item to cart, which firstly displays indication stock level and communicates to the user that they are unable get this product due to low stock.