Order Functionality + Quantity update
Items in cart when pressing checkout actually adds the items to orders. Then, amount ordered will reduce the stock level of the product as part of the brief.

In addition, made quantity maximum to 10 so you cant buy the entire stock level. will adjust to only fit the stock level if stock is less than 10.