Admin Functionality - Order + Remove/Modify Order page
Order page in admin panels, where admins can modify the order_status.

If the order status is modified to delivered, the line will turn green to make it easily readable as order pages will likely be stacked up. Easier visibility. Same with if the order status turns cancelled then that line will turn red.

In addition, added remove button if admins wants to remove the order for whatever reason. Maybe will take that out because cancelled is there already.

As well as, removed unused imports to keep the code cleaner. The order page is using the table that I css'd a while back on carts page so everything is formatted nicely.

Will next make it more easier to read (even more).