Cart Check Validation
If user already has a product in cart and tries to add it again, it will no longer create a separate row from the same product already there because it won't allow duplicates. Quantity does not get updated as well from mass clicking add to cart on a product.

Updated CSS on button and selection on admin orders to ensure it can be seen in dark mode.