Contact Us Functionality + Styling Fixes
Text was hidden on lightmode for contact us form, so that was fixed.

A user can now submit the contact us form which will go the admin feedback page, where admins can view the feedback. Reason to not include user id as a table field for feedback page is that customers that are not logged in can make feedback.