Fixed Products Box CSS + More
Fixed a description problem with add products only allowing 255 characters, so I expanded it to allow descriptions at full length. I then minimised the description that can be seen from the box on the products/home/search pages to make all boxes consistent by having the descriptions be cut off at the same point.

Added categories filtering to the products page.

Fixed global button colourings.

Fixed remove product problem as minimum is an unused field in products table so there was a null violation whenever you attempted to remove a product via the remove button on admin panel.

Will next limit the products to 10 products to be shown on admin panel product page and if the admin wants to see them all, they can do so with 'show all' button.